+++ internal/handlers/auth/validate.go
@@ -1,0 +1,35 @@
+package auth
+
+import (
+	"fmt"
+	"github.com/golang-jwt/jwt/v5"
+	"net/http"
+)
+
+func (h *Handler) Validate(next http.HandlerFunc) http.HandlerFunc {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if len(h.password) > 0 {
+			var jwt string
+			cookie, err := r.Cookie("token")
+			if err == nil {
+				jwt = cookie.Value
+			}
+
+			if !isTokenValid(jwt) {
+				http.Error(w, "Authentification required", http.StatusUnauthorized)
+				return
+			}
+		}
+		next(w, r)
+	})
+}
+
+func isTokenValid(tokenString string) bool {
+	_, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
+		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
+			return nil, fmt.Errorf("Unexpected signing method: %v", token.Header["alg"])
+		}
+		return []byte(secret), nil
+	})
+	return err == nil
+}
