+++ internal/handlers/task_done/post.go
@@ -1,0 +1,67 @@
+package task_done
+
+import (
+	"database/sql"
+	"errors"
+	"net/http"
+	"strconv"
+	"time"
+
+	"go_final_project/internal/models"
+	"go_final_project/internal/utils"
+)
+
+func (h *Handler) handlePostTaskDone(w http.ResponseWriter, r *http.Request) {
+	stringId := r.URL.Query().Get("id")
+	if len(stringId) == 0 {
+		utils.RespondWithError(w, "Не указан идентификатор")
+		return
+	}
+	id, err := strconv.ParseInt(stringId, 10, 64)
+	if err != nil {
+		utils.RespondWithError(w, "Не указан идентификатор")
+		return
+	}
+
+	query := `SELECT 
+    			id,
+    			date,
+    			repeat
+			  FROM scheduler
+			  WHERE id = ?`
+	row := h.db.QueryRow(query, id)
+	var task models.Task
+	err = row.Scan(&task.ID, &task.Date, &task.Repeat)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			utils.RespondWithError(w, "Задача не найдена")
+			return
+		}
+		utils.RespondWithError(w, "Ошибка разбора задач из базы данных")
+		return
+	}
+
+	if len(task.Repeat) == 0 {
+		deleteQuery := `DELETE FROM scheduler WHERE id = ?`
+		_, deleteErr := h.db.Exec(deleteQuery, id)
+		if deleteErr != nil {
+			utils.RespondWithError(w, err.Error())
+			return
+		}
+	} else {
+		task.Date, err = models.NextDate(time.Now(), task.Date, task.Repeat)
+		if err != nil {
+			utils.RespondWithError(w, err.Error())
+			return
+		}
+		updateQuery := `UPDATE scheduler SET date = ? WHERE id = ?`
+		_, err = h.db.Exec(updateQuery, task.Date, task.ID)
+		if err != nil {
+			utils.RespondWithError(w, err.Error())
+			return
+		}
+	}
+
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	w.Write([]byte("{}"))
+}
