+++ internal/models/response.go
@@ -1,0 +1,6 @@
+package models
+
+type Response struct {
+	ID    *int64  `json:"id,omitempty"`
+	Error *string `json:"error,omitempty"`
+}
